+++ CHANGES.md
@@ -11,8 +11,10 @@ Changelog
  - FxApplication provides a Cleaner to use by applications
  - register cleanup actions in FxApplication to execute when application stops
  - FxUtil.copyToClipboard() (supports texts, images, files)
- - Controls.toggleButton()
  - FxUtil.createMenuItem()
+ - FxUtil.dragEventHandler()/dropEventHandler()
+ - Controls.toggleButton()
+ - replace File by Path where possible
 
 #### Version 0.16
 (unreleased)